Add option to use a model from a custom location
This makes it easier for users to download the model separately, especially once they install textwash as a package that does not ship the model in the wheel.

Note that since the model determines the anonymisation language --language and --model_path have been made mutually exclusive. If the user specifies both, they will get an error.